refactor: moved all types to shared folder
