+++ src/actions/category/category.ts
@@ -2,7 +2,7 @@
 import { z } from "zod";
 
 import { db } from "@/shared/lib/db";
-import { TCategory, TGroupJSON } from "@/types/categories";
+import { TCategory, TGroupJSON } from "@/shared/types/categories";
 
 //eslint-disable-next-line
 const GetAllCategories = z.object({
