+++ src/actions/category/categoryOptions.ts
@@ -4,7 +4,7 @@ import { OptionSetType } from "@prisma/client";
 import { z } from "zod";
 
 import { db } from "@/shared/lib/db";
-import { TOptionSet, TSingleOption, TSingleSpec, TSpecGroup } from "@/types/common";
+import { TOptionSet, TSingleOption, TSingleSpec, TSpecGroup } from "@/shared/types/common";
 
 const AddOptionSet = z.object({
   name: z.string().min(3),
