+++ src/actions/pageVisit/pageVisitServices.ts
@@ -4,7 +4,7 @@ import { z } from "zod";
 
 import { TRAFFIC_LIST_PAGE_SIZE } from "@/shared/constants/admin/trafficView";
 import { db } from "@/shared/lib/db";
-import { TAddPageVisit } from "@/types/common";
+import { TAddPageVisit } from "@/shared/types/common";
 
 const ValidatePageVisit = z.object({
   pageType: z.enum(["MAIN", "LIST", "PRODUCT"]),
