+++ src/actions/product/product.ts
@@ -10,7 +10,7 @@ import {
   TProductListItem,
   TProductPageInfo,
   TSpecification,
-} from "@/types/product";
+} from "@/shared/types/product";
 
 const ValidateAddProduct = z.object({
   name: z.string().min(3),
