+++ src/app/admin/products/page.tsx
@@ -7,7 +7,7 @@ import ProductForm from "@/domains/admin/components/product/productForm";
 import ProductListItem from "@/domains/admin/components/product/productListItem";
 import Button from "@/shared/components/UI/button";
 import Popup from "@/shared/components/UI/popup";
-import { TAddProductFormValues, TProductListItem } from "@/types/product";
+import { TAddProductFormValues, TProductListItem } from "@/shared/types/product";
 
 const initialForm: TAddProductFormValues = {
   name: "",
