+++ src/domains/admin/components/category/rowGroup/categoryOptions/AddOption/index.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import { addOptionSet } from "@/actions/category/categoryOptions";
 import Button from "@/shared/components/UI/button";
 import Input from "@/shared/components/UI/input";
-import { TOptionSet } from "@/types/common";
+import { TOptionSet } from "@/shared/types/common";
 
 type TProps = {
   categoryOptionId: string;
