+++ src/domains/admin/components/category/rowGroup/categoryOptions/optionSet/index.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import { addSingleOption, deleteOptionSet, deleteSingleOption } from "@/actions/category/categoryOptions";
 import Button from "@/shared/components/UI/button";
 import Input from "@/shared/components/UI/input";
-import { TOptionSet, TSingleOption } from "@/types/common";
+import { TOptionSet, TSingleOption } from "@/shared/types/common";
 
 type TProps = {
   data: TOptionSet;
