+++ src/domains/admin/components/category/rowGroup/categoryOptions/specGroup/index.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import { addSingleSpec, deleteSingleSpec, deleteSpecGroup } from "@/actions/category/categoryOptions";
 import Button from "@/shared/components/UI/button";
 import Input from "@/shared/components/UI/input";
-import { TSingleSpec, TSpecGroup } from "@/types/common";
+import { TSingleSpec, TSpecGroup } from "@/shared/types/common";
 
 type TProps = {
   data: TSpecGroup;
