+++ src/domains/product/types/dealCards.ts
@@ -1,4 +1,4 @@
-import { TProductCard } from "@/types/common";
+import { TProductCard } from "@/shared/types/common";
 
 export type TDealCard = TProductCard & {
   dealDate: Date;
