+++ src/domains/store/homePage/components/categories/index.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 
 import { getAllCategoriesJSON } from "@/actions/category/category";
 import { SK_Box } from "@/shared/components/UI/skeleton";
-import { TGroupJSON } from "@/types/categories";
+import { TGroupJSON } from "@/shared/types/categories";
 
 import CategoryListItem from "./catListItem";
