+++ src/domains/store/homePage/components/latestBlogPosts/BlogCard.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import { TBlogCard } from "@/types/common";
+import { TBlogCard } from "@/shared/types/common";
 
 const HomeBlogCard = ({ title, imgUrl, shortText, url }: TBlogCard) => {
   return (
