+++ src/shared/types/categories.ts
